style: move shared stuff to utils
