+++ packages/icons/scripts/add-icon.js
@@ -1,33 +1,11 @@
 const fs = require('fs');
 const _ = require('lodash'); // eslint-disable-line lodash/import-scope
 const absolutePath = process.cwd(); // icons package root
-const svgPath = `${absolutePath}/src/svgs`;
+const { getDirectories, getIconDictionary } = require('./utils');
 
-/**
- * Grabs diretories where icons are stored
- * Here object is all the files and directories grabbed from svgPath
- * The endgoal is to only return directories from svgPath
- */
-const directories = fs
-  .readdirSync(svgPath, { withFileTypes: true })
-  .filter((object) => object.isDirectory())
-  .map((object) => object.name);
+const directories = getDirectories();
 
-/**
- * Creates a dictionary for the icons: {us-states: [list of icons], quick-actions: [], etc}
- * It also creates a list of all Icons available for importing
- */
-const allIcons = [];
-let iconDictionary = {};
-
-directories.forEach((directory) => {
-  const arrayOfFiles = fs
-    .readdirSync(`${svgPath}/${directory}`)
-    .filter((name) => !!name.match(/.svg/));
-
-  iconDictionary[directory] = arrayOfFiles;
-  allIcons.push.apply(allIcons, arrayOfFiles);
-});
+const { allIcons, iconDictionary } = getIconDictionary(directories);
 
 const toPascalCase = (someString) => _.startCase(_.camelCase(someString));
